New disks get the 446 bytes of the MS-DOS 6.22 bootcode. It's ugly and hardcoded for now.

+++ src/main.rs
@@ -10,7 +10,7 @@ struct Args {
 
     /// Disk size in bytes
     #[clap(short, long)]
-    size: u32,
+    size: u64,
 }
 fn main() {
     let args = Args::parse();
